requirements.txt
appdirs==1.4.4
asgiref==3.6.0
backports.zoneinfo==0.2.1
buildozer==1.2.0
certifi==2021.5.30
distlib==0.3.0
Django==4.1.7
djangorestframework==3.14.0
filelock==3.0.12
pexpect==4.8.0
pipenv==2021.5.29
ptyprocess==0.6.0
pytz==2022.1
sh==1.13.1
six==1.15.0
sqlparse==0.4.3
tzdata==2022.7
virtualenv==20.0.23
virtualenv-clone==0.5.7
whitenoise==6.3.0
